requirements.txt
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.105.0
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
mako==1.3.10
markupsafe==3.0.2
mypy==1.16.0
mypy-extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
pygments==2.19.1
pytest==8.4.0
pytest-mypy==1.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
ruff==0.11.13
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.23
starlette==0.27.0
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250516
types-python-jose==3.5.0.20250531
typing-extensions==4.14.0
uvicorn==0.24.0
